changes to profile, and some QC logging logic
