+++ src/toolbox/steps/custom/qc/impossible_range_qc.py
@@ -32,7 +32,7 @@ class impossible_range_qc(BaseQC):
     Target Variable: Any
     Flag Number: Any
     Variables Flagged: Any
-    Checks that a meausurement is within a reasonable range.
+    Checks that a measurement is within a reasonable range.
 
     EXAMPLE
     -------
@@ -50,52 +50,43 @@ class impossible_range_qc(BaseQC):
     """
 
     qc_name = "impossible range qc"
-
-    # Specify if test target variable is user-defined (if True, __init__ has to be redefined)
     dynamic = True
 
+    # Define the schema here so BaseQC automatically extracts and assigns them
+    expected_parameters = {
+        "variable_ranges": {},
+        "also_flag": {},
+        "plot": [],
+        "test_depth_range": None
+    }
+
     def __init__(self, data, **kwargs):
-        # Check the necessary kwargs are available
-        required_kwargs = {"variable_ranges", "also_flag", "plot"}
-        if not required_kwargs.issubset(set(kwargs.keys())):
-            raise KeyError(
-                f"{required_kwargs - set(kwargs.keys())} are missing from {self.qc_name} settings"
-            )
+        # 1. Let BaseQC handle data copying, logging setup, and parameter validation
+        super().__init__(data, **kwargs)
 
-        # Specify the tests paramters from kwargs (config)
-        self.expected_parameters = {
-            k: v for k, v in kwargs.items() if k in required_kwargs
-        }
-        self.required_variables = list(
-            set(self.expected_parameters["variable_ranges"].keys())
-        )
-        self.tested_variables = self.required_variables.copy()
-        if "test_depth_range" in kwargs.keys():
+        # 2. Safety check for the required dynamic config
+        if not self.variable_ranges:
+            raise KeyError(f"'variable_ranges' is required but missing from {self.qc_name} settings")
+
+        # 3. Dynamically construct required variables based on user config
+        self.tested_variables = list(self.variable_ranges.keys())
+        self.required_variables = self.tested_variables.copy()
+        
+        if self.test_depth_range is not None:
             self.required_variables.append("DEPTH")
-            self.test_depth_range = kwargs["test_depth_range"]
 
+        # 4. Dynamically construct output columns
         self.qc_outputs = list(
-            set(f"{var}_QC" for var in self.tested_variables)
-            | set(
-                f"{var}_QC"
-                for var in sum(self.expected_parameters["also_flag"].values(), [])
-            )
+            set(f"{var}_QC" for var in self.tested_variables) | 
+            set(f"{var}_QC" for var in sum(self.also_flag.values(), []))
         )
-
-        if data is not None:
-            self.data = data.copy(deep=True)
-
-        for k, v in self.expected_parameters.items():
-            setattr(self, k, v)
 
         self.flags = None
 
     def return_qc(self):
-        # Subset the data
         self.data = self.data[self.required_variables]
 
-        # If the user specified a depth range, limit the checks to that range
-        if hasattr(self, "test_depth_range"):
+        if self.test_depth_range is not None:
             # TODO: -DEPTH
             depth_range_mask = (self.data["DEPTH"] >= self.test_depth_range[0]) & (
                 self.data["DEPTH"] <= self.test_depth_range[1]
@@ -103,14 +94,12 @@ class impossible_range_qc(BaseQC):
         else:
             depth_range_mask = True
 
-        # Make the empty QC columns
         for var in self.tested_variables:
             self.data[f"{var}_QC"] = (
                 ["N_MEASUREMENTS"],
                 np.full(len(self.data[var]), 0),
             )
 
-        # Generate the variable-specific flags
         for var, meta in self.variable_ranges.items():
             for flag, bounds in meta.items():
                 self.data[f"{var}_QC"] = xr.where(
@@ -124,19 +113,16 @@ class impossible_range_qc(BaseQC):
                     0,
                 )
 
-            # Replace all remaining 0s (unchecked) with 1s (good)
             self.data[f"{var}_QC"] = xr.where(
                 depth_range_mask & (self.data[f"{var}_QC"] == 0),
                 1,
                 self.data[f"{var}_QC"],
             )
 
-            # Broadcast the QC found for var into variables specified by "also_flag"
             if extra_vars := self.also_flag.get(var):
                 for extra_var in extra_vars:
                     self.data[f"{extra_var}_QC"] = self.data[f"{var}_QC"]
 
-        # Select just the flags
         self.flags = self.data[
             [var_qc for var_qc in self.data.data_vars if "_QC" in var_qc]
         ]
@@ -146,28 +132,20 @@ class impossible_range_qc(BaseQC):
     def plot_diagnostics(self):
         matplotlib.use("tkagg")
 
-        # If not plots were specified
         if len(self.plot) == 0:
-            print(
-                "WARNING: In 'range test' diagnostics were called but no plots were specified."
-            )
+            self.log_warn("Diagnostics were called but no plots were specified in the config.")
             return
 
-        # Plot the QC output
         fig, axs = plt.subplots(nrows=len(self.plot), figsize=(8, 6), dpi=200)
         if len(self.plot) == 1:
             axs = [axs]
 
         for ax, var in zip(axs, self.plot):
-            # Check that the user specified var exists in the test set
             if f"{var}_QC" not in self.qc_outputs:
-                print(
-                    f"WARNING: Cannot plot {var}_QC as it was not included in this test."
-                )
+                self.log_warn(f"Cannot plot {var}_QC as it was not included in this test.")
                 continue
 
             for i in range(10):
-                # Plot by flag number
                 plot_data = self.data[[var, "N_MEASUREMENTS"]].where(
                     self.data[f"{var}_QC"] == i, drop=True
                 )
@@ -175,7 +153,6 @@ class impossible_range_qc(BaseQC):
                 if len(plot_data[var]) == 0:
                     continue
 
-                # Plot the data
                 ax.plot(
                     plot_data["N_MEASUREMENTS"],
                     plot_data[var],
@@ -185,7 +162,7 @@ class impossible_range_qc(BaseQC):
                     label=f"{i}",
                 )
 
-            for bounds in self.variable_ranges[var].values():
+            for bounds in self.variable_ranges.get(var, {}).values():
                 for bound in bounds:
                     ax.axhline(bound, ls="--", c="k")
 
@@ -194,7 +171,6 @@ class impossible_range_qc(BaseQC):
                 ylabel=var,
                 title=f"{var} Range Test",
             )
-
             ax.legend(title="Flags", loc="upper right")
 
         fig.tight_layout()
